+++ reahl-web/COPYING
@@ -23,10 +23,10 @@ Several files from other projects are included in this code. See the copyright i
 of each applicable file below:
 
 
-reahl-web/reahl/web/static/bootstrap-4.0.0/
+reahl-web/reahl/web/static/bootstrap-4.1.0/
 ===================================================
 
-Bootstrap v4.0.0 (https://getbootstrap.com)
+Bootstrap v4.1.0 (https://getbootstrap.com)
 Copyright 2011-2018 The Bootstrap Authors (https://github.com/twbs/bootstrap/graphs/contributors)
 Licensed under MIT (https://github.com/twbs/bootstrap/blob/master/LICENSE)
