+++ reahl-web/reahl/web_dev/appstructure/expected_hello_world_without_p.html
@@ -7,8 +7,8 @@
         </script><head><title>Hello</title>
 <link rel="stylesheet" href="/static/reahl.css" type="text/css">
 <link rel="stylesheet" href="/static/reahl.runningonbadge.css" type="text/css"><meta charset="utf-8"><meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-<link rel="stylesheet" href="/static/bootstrap-4.0.0/css/bootstrap.css" type="text/css">
-<link rel="stylesheet" href="/static/bootstrap-4.0.0/css/reahl-patch.css" type="text/css">
+<link rel="stylesheet" href="/static/bootstrap-4.1.0/css/bootstrap.css" type="text/css">
+<link rel="stylesheet" href="/static/bootstrap-4.1.0/css/reahl-patch.css" type="text/css">
 <link rel="stylesheet" href="/static/reahl.files.css" type="text/css">
 <link rel="stylesheet" href="/static/reahl.carousel.css" type="text/css">
 <link rel="stylesheet" href="/static/reahl.datatable.css" type="text/css"></head><body>
@@ -42,7 +42,7 @@ $('body').addClass('enhanced');
 <script type="text/javascript" src="/static/reahl.form.js"></script>
 <script type="text/javascript" src="/static/holder-2.9.0.js"></script>
 <script type="text/javascript" src="/static/popper-1.12.9/popper.js"></script>
-<script type="text/javascript" src="/static/bootstrap-4.0.0/js/bootstrap.js"></script>
+<script type="text/javascript" src="/static/bootstrap-4.1.0/js/bootstrap.js"></script>
 <script type="text/javascript" src="/static/reahl.bootstrapform.js"></script>
 <script type="text/javascript" src="/static/reahl.pagination.js"></script>
 <script type="text/javascript" src="/static/reahl.files.js"></script>
